Refactoring project structure
Moving to a modular structure rather than hierarchical. With the modules
we can pick what services we want to include in each of images, and so
it will facilitate building out test images. We can create new modules
for new tests and then simply add them to our images.

The modules have everything set-up so that all the image has to do is
include it in the mkosi.conf file.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>