feat: Adding services to prepare sev-snp enabled guest
Added modules that will be needed to succesfully launch snpguest for
testing. Those include a service to check the system is set up correctly
with snphost ok, calculate the expected measurement with snpguest and
then actually building out and using a qemu command line that will use
the embeded guest image.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>