Refactor: Updating log collection for guest tests in host
A few changes to improve log collection in the host:

- Removed the show-snp-guest-attestation module and replaced it with
  ./logging/display-guest-logs
- ./logging/display-guest-logs simply waits for tests to be finished
  running on the guest and display the results. There is no need to
communicate failures when displaying guest logs, since the tests already
communicate that.
- ./logging/display-guest-logs will also check to see if the capture
  logs services is working correctly.
- Added a new module called guest-notices. This is 2 simple services to
  log that the boot was succesful and that the tests finished running.
This will simplify checking for success in the host.
- Added a separate service called verify-guest. This service will check
  to see if the guest finished booting. The launch-guest service will
only attempt to launch the guest and keep it up, giving the guest
enought time to finish running the tests, also allows users to access
the guest in case of debbuging.
- read-guest-journal is replaced by display-guest-logs.
- in logging renamed guest to upload-guest-logs
- Created a snpguest-tests.target that can be modified to add more
  tests. The target can be used to let us know that the tests have
finished running instead of having to check each test individually.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>